+++ srcs/stack/stack_b_data.c
@@ -6,7 +6,7 @@
 /*   By: abahoumi <abahoumi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/26 15:42:09 by abahoumi          #+#    #+#             */
-/*   Updated: 2025/11/28 10:15:53 by abahoumi         ###   ########.fr       */
+/*   Updated: 2025/12/01 12:12:14 by abahoumi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@ static void	set_target_b(t_stack *a, t_stack *b)
 {
 	t_stack	*target_node;
 	t_stack	*current_a;
-	int		current_target;
+	long	current_target;
 
 	while (b)
 	{
 		current_a = a;
-		current_target = INT_MAX;
+		current_target = LONG_MAX;
 		while (current_a)
 		{
 			if (b->value < current_a->value
@@ -32,7 +32,7 @@ static void	set_target_b(t_stack *a, t_stack *b)
 			}
 			current_a = current_a->next;
 		}
-		if (current_target == INT_MAX)
+		if (current_target == LONG_MAX)
 			b->target_node = find_smallest(a);
 		else
 			b->target_node = target_node;
